Start access check in content provider

+++ sensappdroid/src/main/java/org/sensapp/android/sensappdroid/contentprovider/SensorCP.java
@@ -18,36 +18,42 @@ import android.database.sqlite.SQLiteQueryBuilder;
 import android.net.Uri;
 import android.text.TextUtils;
 
-public class SensorCP {
+public class SensorCP extends TableContentProvider {
 	
 	protected static final String BASE_PATH = "sensors";
 	
 	private static final int SENSORS = 10;
 	private static final int COMPOSITE_SENSORS = 20;
-	private static final int SENSOR_ID = 30;
-	private static final UriMatcher measureURIMatcher = new UriMatcher(UriMatcher.NO_MATCH);
+	private static final int SENSOR_UID = 30;
+	private static final int SENSOR_ID = 40;
+	
+	private static final UriMatcher sensorURIMatcher = new UriMatcher(UriMatcher.NO_MATCH);
 	static {
-		measureURIMatcher.addURI(SensAppContentProvider.AUTHORITY, BASE_PATH, SENSORS);
-		measureURIMatcher.addURI(SensAppContentProvider.AUTHORITY, BASE_PATH + "/composite/*", COMPOSITE_SENSORS);
-		measureURIMatcher.addURI(SensAppContentProvider.AUTHORITY, BASE_PATH + "/*", SENSOR_ID);
+		sensorURIMatcher.addURI(SensAppContentProvider.AUTHORITY, BASE_PATH, SENSORS);
+		sensorURIMatcher.addURI(SensAppContentProvider.AUTHORITY, BASE_PATH + "/composite/*", COMPOSITE_SENSORS);
+		sensorURIMatcher.addURI(SensAppContentProvider.AUTHORITY, BASE_PATH + "/uid/*", SENSOR_UID);
+		sensorURIMatcher.addURI(SensAppContentProvider.AUTHORITY, BASE_PATH + "/*", SENSOR_ID);
 	}
-
-	private SensAppDatabaseHelper database;
-	private Context context;
 	
 	public SensorCP(Context context, SensAppDatabaseHelper database) {
-		this.context = context;
-		this.database = database;
+		super(context, database);
 	}
 	
-	public Cursor query(Uri uri, String[] projection, String selection, String[] selectionArgs, String sortOrder) {
+	@Override
+	public Cursor query(Uri uri, String[] projection, String selection, String[] selectionArgs, String sortOrder, int uid) throws IllegalStateException {
 		checkColumns(projection);
 		SQLiteQueryBuilder queryBuilder = new SQLiteQueryBuilder();
-		switch (measureURIMatcher.match(uri)) {
+		switch (sensorURIMatcher.match(uri)) {
 		case SENSORS:
+			if (!isSensAppUID(uid)) {
+				throw new IllegalStateException("Forbiden uri");
+			}
 			queryBuilder.setTables(SensorTable.TABLE_SENSOR);
 			break;
 		case COMPOSITE_SENSORS:
+			if (!isSensAppUID(uid)) {
+				throw new IllegalStateException("Forbiden uri");
+			}
 			Hashtable<String, String> columnMap = new Hashtable<String, String>();
 			columnMap.put(SensorTable.COLUMN_NAME, SensorTable.TABLE_SENSOR + "." + SensorTable.COLUMN_NAME);
 			columnMap.put(SensorTable.COLUMN_BACKEND, SensorTable.TABLE_SENSOR + "." + SensorTable.COLUMN_BACKEND);
@@ -61,89 +67,112 @@ public class SensorCP {
 			queryBuilder.appendWhere(SensorTable.TABLE_SENSOR + "." + SensorTable.COLUMN_NAME + " = " + ComposeTable.COLUMN_SENSOR
 					 + " AND " + ComposeTable.COLUMN_COMPOSITE + " = \"" + uri.getLastPathSegment() + "\"");
 			break;
+		case SENSOR_UID:
+			queryBuilder.setTables(SensorTable.TABLE_SENSOR);
+			queryBuilder.appendWhere(SensorTable.COLUMN_NAME + " = \"" + uri.getLastPathSegment() + "\"");
+			projection = new String[]{SensorTable.COLUMN_CLIENT_UID};
+			selection = null;
+			break;
 		case SENSOR_ID:
+			int sensorUID = getSensorUID(uri.getLastPathSegment());
+			if (!isSensAppUID(uid) && sensorUID != 0 && sensorUID != uid) {
+				throw new IllegalStateException("Forbiden uri");
+			}
 			queryBuilder.setTables(SensorTable.TABLE_SENSOR);
 			queryBuilder.appendWhere(SensorTable.COLUMN_NAME + " = \"" + uri.getLastPathSegment() + "\"");
 			break;
 		default:
 			throw new IllegalArgumentException("Unknown URI: " + uri);
 		}
-		SQLiteDatabase db = database.getWritableDatabase();
+		SQLiteDatabase db = getDatabase().getWritableDatabase();
 		Cursor cursor = queryBuilder.query(db, projection, selection, selectionArgs, null, null, sortOrder);
-		cursor.setNotificationUri(context.getContentResolver(), uri);
+		cursor.setNotificationUri(getContext().getContentResolver(), uri);
 		return cursor;
 	}
-
-	public String getType(Uri uri) {
-		return null;
-	}
 	
-	public Uri insert(Uri uri, ContentValues values) {
-		SQLiteDatabase db = database.getWritableDatabase();
+	@Override
+	public Uri insert(Uri uri, ContentValues values, int uid) {
+		SQLiteDatabase db = getDatabase().getWritableDatabase();
 		long id = 0;
-		switch (measureURIMatcher.match(uri)) {
+		switch (sensorURIMatcher.match(uri)) {
 		case SENSORS:
 			values.put(SensorTable.COLUMN_UPLOADED, 0);
+			values.put(SensorTable.COLUMN_CLIENT_UID, uid);
 			id = db.insert(SensorTable.TABLE_SENSOR, null, values);
 			break;
 		default:
 			throw new IllegalArgumentException("Bad URI: " + uri);
 		}
-		context.getContentResolver().notifyChange(uri, null);
-		return Uri.parse(BASE_PATH + "/" + id);
+		getContext().getContentResolver().notifyChange(uri, null);
+		return Uri.parse("content://" + SensAppContentProvider.AUTHORITY + "/" + BASE_PATH + "/" + id);
 	}
 	
-	public int delete(Uri uri, String selection, String[] selectionArgs) {
-		SQLiteDatabase db = database.getWritableDatabase();
+	@Override
+	public int delete(Uri uri, String selection, String[] selectionArgs, int uid) throws IllegalStateException {
+		SQLiteDatabase db = getDatabase().getWritableDatabase();
 		int rowsDeleted = 0;
-		switch (measureURIMatcher.match(uri)) {
+		switch (sensorURIMatcher.match(uri)) {
 		case SENSORS:
+			if (!isSensAppUID(uid)) {
+				throw new IllegalStateException("Forbiden uri");
+			}
 			rowsDeleted = db.delete(SensorTable.TABLE_SENSOR, selection, selectionArgs);
 			// Clean composites
 			db.delete(ComposeTable.TABLE_COMPOSE, null, null);
 			break;
 		case SENSOR_ID:
-			String id = uri.getLastPathSegment();
+			String name = uri.getLastPathSegment();
+			if (!isSensAppUID(uid) && getSensorUID(name) != uid) {
+				throw new IllegalStateException("Forbiden uri");
+			}
 			if (TextUtils.isEmpty(selection)) {
-				rowsDeleted = db.delete(SensorTable.TABLE_SENSOR, SensorTable.COLUMN_NAME + " = \"" + id + "\"", null);
+				rowsDeleted = db.delete(SensorTable.TABLE_SENSOR, SensorTable.COLUMN_NAME + " = \"" + name + "\"", null);
 			} else {
-				rowsDeleted = db.delete(SensorTable.TABLE_SENSOR, SensorTable.COLUMN_NAME + " = \"" + id + "\" and " + selection, selectionArgs);
+				rowsDeleted = db.delete(SensorTable.TABLE_SENSOR, SensorTable.COLUMN_NAME + " = \"" + name + "\" and " + selection, selectionArgs);
 			}
 			// Clean composites
-			db.delete(ComposeTable.TABLE_COMPOSE, ComposeTable.COLUMN_SENSOR + " = \"" + id + "\"", null);
+			db.delete(ComposeTable.TABLE_COMPOSE, ComposeTable.COLUMN_SENSOR + " = \"" + name + "\"", null);
 			break;
 		default:
 			throw new IllegalArgumentException("Unknown URI: " + uri);
 		}
-		context.getContentResolver().notifyChange(uri, null);
-		context.getContentResolver().notifyChange(Uri.parse("content://" + SensAppContentProvider.AUTHORITY + "/" + BASE_PATH + "/composite"), null);
+		getContext().getContentResolver().notifyChange(uri, null);
+		getContext().getContentResolver().notifyChange(Uri.parse("content://" + SensAppContentProvider.AUTHORITY + "/" + BASE_PATH + "/composite"), null);
 		return rowsDeleted;
 	}
 
-	public int update(Uri uri, ContentValues values, String selection, String[] selectionArgs) {
-		SQLiteDatabase db = database.getWritableDatabase();
+	@Override
+	public int update(Uri uri, ContentValues values, String selection, String[] selectionArgs, int uid) throws IllegalStateException {
+		SQLiteDatabase db = getDatabase().getWritableDatabase();
 		int rowsUpdated = 0;
-		switch (measureURIMatcher.match(uri)) {
+		switch (sensorURIMatcher.match(uri)) {
 		case SENSORS:
+			if (!isSensAppUID(uid)) {
+				throw new IllegalStateException("Forbiden uri");
+			}
 			rowsUpdated = db.update(SensorTable.TABLE_SENSOR, values, selection, selectionArgs);
 			break;
 		case SENSOR_ID:
-			String id = uri.getLastPathSegment();
+			String name = uri.getLastPathSegment();
+			if (!isSensAppUID(uid) && getSensorUID(name) != uid) {
+				throw new IllegalStateException("Forbiden uri");
+			}
 			if (TextUtils.isEmpty(selection)) {
-				rowsUpdated = db.update(SensorTable.TABLE_SENSOR, values, SensorTable.COLUMN_NAME + " = \"" + id + "\"", null);
+				rowsUpdated = db.update(SensorTable.TABLE_SENSOR, values, SensorTable.COLUMN_NAME + " = \"" + name + "\"", null);
 			} else {
-				rowsUpdated = db.update(SensorTable.TABLE_SENSOR, values, SensorTable.COLUMN_NAME + " = \"" + id + "\" and " + selection, selectionArgs);
+				rowsUpdated = db.update(SensorTable.TABLE_SENSOR, values, SensorTable.COLUMN_NAME + " = \"" + name + "\" and " + selection, selectionArgs);
 			}
 			break;
 		default:
 			throw new IllegalArgumentException("Unknown URI: " + uri);
 		}
-		context.getContentResolver().notifyChange(uri, null);
+		getContext().getContentResolver().notifyChange(uri, null);
 		return rowsUpdated;
 	}
 	
-	private void checkColumns(String[] projection) {
-		String[] available = {SensorTable.COLUMN_NAME, SensorTable.COLUMN_URI, SensorTable.COLUMN_DESCRIPTION,SensorTable.COLUMN_BACKEND, SensorTable.COLUMN_UNIT, SensorTable.COLUMN_TEMPLATE, SensorTable.COLUMN_UPLOADED, SensorTable.COLUMN_ICON};
+	@Override
+	protected void checkColumns(String[] projection) {
+		String[] available = {SensorTable.COLUMN_NAME, SensorTable.COLUMN_URI, SensorTable.COLUMN_DESCRIPTION,SensorTable.COLUMN_BACKEND, SensorTable.COLUMN_UNIT, SensorTable.COLUMN_TEMPLATE, SensorTable.COLUMN_UPLOADED, SensorTable.COLUMN_ICON, SensorTable.COLUMN_CLIENT_UID};
 		if (projection != null) {
 			HashSet<String> requestedColumns = new HashSet<String>(Arrays.asList(projection));
 			HashSet<String> availableColumns = new HashSet<String>(Arrays.asList(available));
